FIX: Improve the logic to rotate error log files
On certain environment (e.g. embedded systems), the error log is saved in
flash or memory and limiting the space for error log is needed.

The current implementation checks whether the log file is larger than
the configured size in every 1000 messages. Until then the error log
keeps growing, which does not control the log file size well enough.

For example, if log size is configured with the minimum (10KB) and average log
message is 100 bytes, then the log file will grow up to 100KB (100 * 1000) until
attemping to back up the log.

So to keep the log file size close to the configured limit, accumulate
the number of bytes written and call backupLog() if it is over the limit.